Fix a bug in SetAuthorizations
HW_AUTH_PASSWORD and HW_AUTH_FINGERPRINT are bit masks
and should be used as such when they are present in
in USER_AUTH_TYPE Tag.

Current implementation of Trusty does support TEE based
enforcement of HW_AUTH_PASSWORD but does not support
HW_AUTH_FINGERPRINT so we cannot advertise that it is
TEE enforced.

If system integrators add support for handling
HW_AUTH_FINGERPRINT type in Trusty they need to set
TEE_FINGERPRINT_AUTH_SUPPORTED to enable it's handling.

Bug: 78457849
Change-Id: I1e98b382cca74a70dc7acf443d25a8ca520bf66c