Check in prebuilt protobuf file to fix build issue.
The nanopb protobuf file building tool chain would not work in some
environment with a lower version of protobuf. The lower version would
be picked instead of the higher version that is checked in the tree.
We currently don't have a good way to solve this problem so we check
in the prebuilt protobuf file into the tree as a workaround.

Bug: 120168945
Test: Build pass.

Change-Id: I30d33f476d55ab27fc3e710f3195b2e9a282e866