Add special handling for AES keys with TAG_STORAGE_KEY tag
This CL adds an optional special handing for for AES keys with
TAG_STORAGE_KEY. This facility is to support hardware specific
storage encryption infrastructure and operates in conjunction
with an external hardware specific IPC service that implements
HWWSK (Hardware wrapped storage key) protocol which provides
backend support.

Once enabled (requires WITH_HWWSK_SUPPORT #define set), the
request to generate/import such keys will be redirected to HWWSK
service for handling which should produce opaque HWWSK key blob.
This blob, in turn, is encrypted as a normal keymaster key and
returned to the caller. The resulting keymaster key can be only
usable for export operation which, when invoked, is also forwarded
for processing to HWWSK service and resulting blob is returned to
caller.

Bug: None
Change-Id: I37506f2214ccad7ed08abcfa19cfe840a518551b