Handle keystore2 blob prefixes
The keystore2 km_compat layer that wraps Keymaster hardware
implementations prefixes the Keymaster blobs with some magic and an
indicator value.  When Keymaster is upgraded in-place to KeyMint,
keystore2 will stop using km_compat and pass the prefixed blobs to
KeyMint as-is.  This change causes Trusty KeyMint to look for and
strip those prefixes.

Bug: 177729159
Test: Manually create keys, update HAL and verify keys work
Change-Id: I5a7f322206bce1c5a384666fa3200afaa62f1f5a