Correct handling of keystore-prefixed blobs.
keystore2 adds some blob prefixes to TrustyKeymaster key blobs, with a
byte to indicate whether the blob is hardware-backed or a software
blob.  The Trusty KeymasterContext had the values of this indicator
byte inverted, causing Trusty KeyMint to interpret software blobs as
hardware and vice versa, making it incapable of using either.

Bug: 177729159
Test: VtsAidlKeyMintTargetTest
Change-Id: I26083e7d375c9dc85966181508c3611eadc16c14