Trusty implementation for IRemotelyProvisi... HAL
This is the initial change for the IRemotelyProvisionedComponent HAL
implementation for Trusty. This is still missing a proper BCC generation
method based off of the underlying hardware keys. It provides a context
implementation that extends RemoteProvisioningContext, along with
accompanying message forwarding in the ipc class for the two functions
defined in the HAL interface.

Additionally, there is light refactoring of the boot parameters in
trusty_keymaster_context. The various parameters were placed into a
struct to make passing them into the remote provisioning context easier.

Bug: 192228022
Test: atest VtsHalRemotelyProvisionedComponentTargetTest
Change-Id: I7fa171997e8eee6e34f95fd40f0a0eebfb5bd78f