Adding an API for reading/writing Attestation IDs
Trusty needs to be able to have Attestation IDs (product, manufacturer,
device, IMEI, etc) provisioned into secure storage in order to be KM
compliant. Device IDs should only be attested if the values passed in
from frameworks matches the values provisioned into the TEE, or not at
all if destroyAttestationIds() is called on the KM HAL/AIDL interface.

This commit introduces the mechanism for storing and reading those
device IDs. Notably, the IDs can only be provisioned once unless
KEYMASTER_DEBUG is set to true.

Bug: 178796950
Test: VtsHalKeymasterV4_0TargetTest
Change-Id: I05737996514c0ac2f67d6492bb516305519eddb4