Use the Trusty hwbcc service for RKP
This change replaces the in-app computations for generating the BCC and
signing the MAC key with the equivalent calls into the hardware backend.
This future proofs the design for stage two of RKP and later, when the
hardware will necessarily hold the boot certificate chain and the device
private key.

Bug: 194509629
Test: atest VtsHalRemotelyProvisionedComponentTest
Change-Id: Icebb37d75b6d6f696a2026f719e093d3f0615d17