Trusty KeyMint: add missing bootloader message
Previous commit 1ea9175c38c6 ("Trusty KeyMint: support boot,vendor
patchlevels") added a new message that allows the bootloader to
send the boot patchlevel to KeyMint.  However, the message is
rejected on receipt because KeyMint hasn't been configured at the time
that it arrives.

Add the new message to cmd_is_from_bootloader() to allow early
processing of this message.

Bug: 193423844
Test: VtsAidlKeyMintTargetTest
Change-Id: I0ff2f2293737be777ee620d34db881d286f86d09