Enable UpgradeKey to add rollback protection.
This is an obscure case, and one not mentioned in the specification,
but we may want to be able to add rollback protection when we upgrade
a key blob.  This will not magically make the key securely-deletable,
because as long as someone can find the pre-upgrade blob they can use
the key, but assuming that pre-upgrade blob is eventually lost, it's
useful to make this upgrade.

Bug: 187105270
Test: VtsAidlKeyMintTargetTest
Change-Id: I8799eed941a9a2a970ce1da37429f8ebbc346935