MA-19722 keymaster: disable secure deletion
After enaling secure deletion, "factory reset" secret and
"secure deletion" secret would be stored in the rpmb storage.
But after enaling the secure deletion, device reboot would
fail after setting PIN/Pattern code for multi-user. The
story is:
  1. Device boot with Owner, Keymint service starts at
     'late-fs' stage. Create new keys which are bound to
     the Owner, these keys are not stored into RPMB as
     secure storage is not ready.
  2. Secure storage starts at 'post-fs-data' stage. New keys
     can be stored into the RPMB storage.
  3. Device boot up with Owner.
  4. Create Guest user and set PIN code for it. Create new
     keys which are bound to the Guest. These keys are stored
     into RPMB as secure storage is ready.
  5. Reboot the device. Keymint sevice starts and will load
     the keys from RPMB for the Guest. But as the secure storage
     is not ready at this stage, Keymint will always wait and
     block the boot process.

As the TD port in secure storage service will depend on the
userdata, so we can only start the secure storage when userdata
is ready at 'post-fs-data'.

This commit disables the secure deletion before we correctly
handle the TD port and the userdata partition.

Signed-off-by: Ji Luo <ji.luo@nxp.com>
Change-Id: I46ce0629d9b7ce9d3c998daa29eaabf8247da6f0