MA-19771 Use testkey for checking confirmation token
As we are using the powerkey event as the 'NotSoSecureInput'
input method, test key is used as hmac key now. Here use the
test key as hmac key to check the confirmation token to make
the Android Protected Confirmation cts verifier pass.

We need to revert this patch once the secure input and true
hmac key are implemented.

Test: cts verifier.

Signed-off-by: Ji Luo <ji.luo@nxp.com>
Change-Id: Ibf5fd59367309505a4233f08c6238c948ad1f84a