Remove undef STRINGIFY
This appears to have been added because of a conflict between the
keymaster logger.h header and the unit test macros.

The keymaster STRINGIFY macros have been renamed to avoid namespace
conflicts, so the undef STRINGIFY in this code is no longer necessary.
In fact, if we leave the compile breaks because it's assumed that
STRINGIFY is defined multiple times.

Test: host-test:keymaster_test
Change-Id: I8d02ff08bf12826477a87cc15964e8d95e779ce5