Add IPC port to manifest
If the port for an application is not available, ql-tipc
immediately returns with an error without waiting for the port
unless that port is listed in the manifest. This adds the Keymaster
port to the manifest to fix a race between the bootloader
connecting to Keymaster and the app starting during boot.

Bug: 217580763
Change-Id: Ica2962e2962e05790c4f9b230165118c92dfac42